refactor: separate file function into util
